requirements.txt
streamlit>=1.28.0
jinja2>=3.1.0
weasyprint>=60.0
python-dotenv>=1.0.0
langchain-openai>=0.0.5
pydantic>=2.0.0
pytest>=7.0.0